move OfflineBar to a new file BreathLoadingBar manages its height constraint BreathLoadingBar has its state for handling animation
